util: rename uki-util.[ch] -> kernel-image.[ch]
As the file contains functions not only for UKI.
