+++ src/shared/kernel-image.h
