+++ src/shared/meson.build
@@ -91,6 +91,7 @@ shared_sources = files(
         'journal-util.c',
         'json.c',
         'kbd-util.c',
+        'kernel-image.c',
         'keyring-util.c',
         'killall.c',
         'label.c',
@@ -150,7 +151,6 @@ shared_sources = files(
         'tpm2-util.c',
         'udev-util.c',
         'uid-alloc-range.c',
-        'uki-util.c',
         'user-record-nss.c',
         'user-record-show.c',
         'user-record.c',
